ARC-2412: Workers should use encrypted connection to database (#112)
Refactored the connection code to a separate file and updated labeeb app and workers code to use the same code path to connect to the database.
